requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic-settings>=2.0
anthropic>=0.49
httpx>=0.27
python-dotenv>=1.0
